Abstract api.rs.
